(maint) Add a better example and tests
The example in the docs for extra_config_section of an array of hashes
is quite confusing, as it leads the reader to believe that the best way
of implementing refresh_patterns is via extra_config_section.

Here, I replace this with a more realistic example using always_direct.

Further, I add a passing unit test case for the array of hashes case.