Clean up module design
This is a breaking change.

What is changing:

- (breaking) Remove defaults for parameters access_log, cache_mem and
  maximum_object_size_in_memory. Also, these parameters have been declared
  as the Optional data type.

- (breaking) Remove boolean support from parameter
  maximum_object_size_in_memory. Previously it accepted 'on', 'off'
  as well as booleans.

- (breaking) Minor reformatting in templates/squid.conf.header.erb, to
  clean up indentation, remove blank lines, and sort parameters
  alphabetically.

- Parameters sorted alphabetically in init.pp, config.pp and README.md.

- Unit tests updated.

Why is this changing:

Without this patch applied, the design is over-complicated and confusing
in the way that Squid's built-in defaults are respected in all contexts
other than the access_log, cache_mem and maximum_object_size_in_memory.
Furthermore, the decision to allow booleans in
maximum_object_size_in_memory seems to violate the keep-it-simple
principle.

As I plan to add quite a few new features, it seems to make sense to
clarify the design implicit in the code before doing so.

What to expect when upgrading:

- The location of the access_log is likely to move to the Squid default
  location. To avoid this, find out what it was before and explicitly
  pass this value to the access_log parameter.

- If you were passing in a boolean to maximum_object_size_in_memory, you
  will need to change the code to pass in 'on' or 'off'.

- If you were relying on the default values for cache_mem and/or
  maximum_object_size_in_memory, there should be a change to the
  squid.conf file after this change, but no actual change to the behaviour
  of Squid, since the Puppet module matched Squid's own defaults.